+++ lib/asset_cache.dart
@@ -1,3 +1,8 @@
 library asset_cache;
 
+export 'src/generic_cache.dart';
 export 'src/asset_cache.dart';
+export 'src/assets_image.dart';
+export 'src/assets_byte_data.dart';
+export 'src/assets_json.dart';
+export 'src/assets_string.dart';