+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [0.2.0]
+- split up asset_cache in multiple files
+- changed Uint8List decoder class to use ByteData
+- updated README
+
 ## [0.1.4]
 - update README | format